[AIRFLOW-5786] Migrate AWS SNS to /providers/amazon/aws (#6502)
* [AIRFLOW-5786] Move AWS Sns hook and operator according to AIP-21
