+++ docs/autoapi_templates/index.rst
@@ -76,6 +76,10 @@ All operators are in the following packages:
 
   airflow/providers/amazon/aws/sensors/index
 
+  airflow/providers/amazon/aws/operators/index
+
+  airflow/providers/amazon/aws/sensors/index
+
   airflow/providers/google/cloud/operators/index
 
   airflow/providers/google/marketing_platform/operators/index
@@ -105,6 +109,8 @@ All hooks are in the following packages:
   airflow/contrib/hooks/index
 
   airflow/gcp/hooks/index
+
+  airflow/providers/amazon/aws/hooks/index
 
   airflow/providers/amazon/aws/hooks/index
