+++ docs/conf.py
@@ -220,6 +220,9 @@ exclude_patterns = [
     '_api/airflow/providers/google/marketing_platform/example_dags',
     '_api/airflow/providers/google/cloud/index.rst',
     '_api/airflow/providers/google/cloud/example_dags',
+    '_api/airflow/providers/amazon/index.rst',
+    '_api/airflow/providers/amazon/aws/index.rst',
+    '_api/airflow/providers/amazon/aws/example_dags',
     'autoapi_templates',
     'howto/operator/gcp/_partials',
 ]
