+++ tests/test_core_to_contrib.py
@@ -144,6 +144,10 @@ HOOK = [
         "airflow.providers.amazon.aws.hooks.lambda_function.AwsLambdaHook",
         "airflow.contrib.hooks.aws_lambda_hook.AwsLambdaHook",
     ),
+    (
+        "airflow.providers.amazon.aws.hooks.sns.AwsSnsHook",
+        "airflow.contrib.hooks.aws_sns_hook.AwsSnsHook",
+    ),
 ]
 OPERATOR = [
     (
@@ -757,6 +761,10 @@ OPERATOR = [
     (
         "airflow.providers.amazon.aws.operators.sqs.SQSPublishOperator",
         "airflow.contrib.operators.aws_sqs_publish_operator.SQSPublishOperator",
+    ),
+    (
+        "airflow.providers.amazon.aws.operators.sns.SnsPublishOperator",
+        "airflow.contrib.operators.sns_publish_operator.SnsPublishOperator",
     )
 ]
 SENSOR = [
